+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add more Slack bots.
 - Handle instagram user agents that have a slash instead of a space.
+- Promote Snapchat to a browser (it was detected as a bot previously).
 
 ## 2.6.1